# openapi-Swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /pagamento/{id}:
    get:
      tags:
        - pagamento-controller
      operationId: getById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    put:
      tags:
        - pagamento-controller
      operationId: atualizaPagamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pagamento'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    delete:
      tags:
        - pagamento-controller
      operationId: deletaPagamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
...
  /pagamento:
    get:
      tags:
        - pagamento-controller
      operationId: getAll_4
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pagamento'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    post:
      tags:
        - pagamento-controller
      operationId: criarPagamento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PagamentoDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Pagamento'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
...
components:
  schemas:
    Avaliacao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        titulo:
          type: string
        classificacao:
          type: integer
          format: int32
        dataAvaliacao:
          type: string
          format: date
        comentario:
          type: string
        passeio:
          $ref: '#/components/schemas/Pacote'
        cliente:
          $ref: '#/components/schemas/Person'
    Pacote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nome:
          type: string
        descricao:
          type: string
        imagem:
          type: string
        ativo:
          type: boolean
        valor:
          type: number
          format: double
        origem:
          type: string
        destino:
          type: string
        itinerario:
          type: string
        dataHoraSaida:
          type: string
        dataHoraRetorno:
          type: string
        quantidadeMaximaPessoas:
          type: integer
          format: int32
        observacao:
          type: string
        reservas:
          type: array
          items:
            $ref: '#/components/schemas/Reserva'
        avaliacoes:
          type: array
          items:
            $ref: '#/components/schemas/Avaliacao'
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nome:
          type: string
        email:
          type: string
        password:
          type: string
        cpf:
          type: string
        cnpj:
          type: string
        dataNascimento:
          type: string
          format: date-time
        telefone:
          type: string
        cliente:
          type: boolean
        reservas:
          type: array
          items:
            $ref: '#/components/schemas/Reserva'
        avaliacoes:
          type: array
          items:
            $ref: '#/components/schemas/Avaliacao'
    Reserva:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        person:
          $ref: '#/components/schemas/Person'
        pacote:
          $ref: '#/components/schemas/Pacote'
        dataReserva:
          type: string
          format: date
        qtdPessoas:
          type: integer
          format: int32
...
    Pagamento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        valor:
          type: number
          format: double
        dataPagamento:
          type: string
          format: date-time
        status:
          type: string
        formaPagamento:
          type: string
        reserva:
          $ref: '#/components/schemas/Reserva'
...
    PagamentoDTO:
      required:
        - dataPagamento
        - formaPagamento
        - pessoa
        - reserva
        - status
        - valor
      type: object
      properties:
        valor:
          type: number
          format: double
        dataPagamento:
          type: string
          format: date-time
        status:
          type: string
        formaPagamento:
          type: string
        reserva:
          $ref: '#/components/schemas/Reserva'
        pessoa:
          $ref: '#/components/schemas/Person'
...